requirements.txt
joblib>=1.1.0,<2.0.0
boto3
snowflake-snowpark-python[pandas]
git+https://gitlab.com/vopak/datascience/vopak_ds_library@v04.05.00
requests
pytz
plotly
streamlit
